NikonLASDecompressor: 'val' is just a bitfield. keep it unsigned.
Same issue as with the normal huffman table implementation.
I didn't yet actually flip the storage type of 'val' though.

../src/librawspeed/decompressors/NikonDecompressor.cpp:263:30: runtime error: implicit conversion from type 'unsigned int' of value 4294959368 (32-bit, unsigned) to type '__gnu_cxx::__alloc_traits<std::allocator<int>, int>::value_type' (aka 'int') changed the value to -7928 (32-bit, signed)
    #0 0x6809c9 in rawspeed::(anonymous namespace)::NikonLASDecompressor::createHuffmanTable() /home/lebedevri/rawspeed/build-Clang-SANITIZE/../src/librawspeed/decompressors/NikonDecompressor.cpp
    #1 0x6809c9 in rawspeed::(anonymous namespace)::NikonLASDecompressor::setup(bool, bool) /home/lebedevri/rawspeed/build-Clang-SANITIZE/../src/librawspeed/decompressors/NikonDecompressor.cpp:286
    #2 0x6809c9 in rawspeed::(anonymous namespace)::NikonLASDecompressor rawspeed::NikonDecompressor::createHuffmanTable<rawspeed::(anonymous namespace)::NikonLASDecompressor>(unsigned int) /home/lebedevri/rawspeed/build-Clang-SANITIZE/../src/librawspeed/decompressors/NikonDecompressor.cpp:429
    #3 0x67da3c in void rawspeed::NikonDecompressor::decompress<rawspeed::(anonymous namespace)::NikonLASDecompressor>(rawspeed::BitStream<rawspeed::MSBBitPumpTag, rawspeed::BitStreamCacheRightInLeftOut>*, int, int) /home/lebedevri/rawspeed/build-Clang-SANITIZE/../src/librawspeed/decompressors/NikonDecompressor.cpp:480:16
    #4 0x67da3c in rawspeed::NikonDecompressor::decompress(rawspeed::ByteStream const&, bool) /home/lebedevri/rawspeed/build-Clang-SANITIZE/../src/librawspeed/decompressors/NikonDecompressor.cpp:538
    #5 0x666a4e in rawspeed::NefDecoder::decodeRawInternal() /home/lebedevri/rawspeed/build-Clang-SANITIZE/../src/librawspeed/decoders/NefDecoder.cpp:125:5
    #6 0x6c13dd in rawspeed::RawDecoder::decodeRaw() /home/lebedevri/rawspeed/build-Clang-SANITIZE/../src/librawspeed/decoders/RawDecoder.cpp:260:20
    #7 0x4c58a5 in rawspeed::rstest::process(std::__cxx11::basic_string<char, std::char_traits<char>, std::allocator<char> > const&, rawspeed::CameraMetaData const*, rawspeed::rstest::options const&) /home/lebedevri/rawspeed/build-Clang-SANITIZE/../src/utilities/rstest/rstest.cpp:380:12

+++ src/librawspeed/decompressors/NikonDecompressor.cpp
@@ -260,7 +260,8 @@ class NikonLASDecompressor {
         int x = input >> (16 - l - rv) & ((1 << rv) - 1);
         if ((x & (1 << (rv - 1))) == 0)
           x -= (1 << rv) - 1;
-        dctbl1.bigTable[i] = (static_cast<unsigned>(x) << 8) | (l + rv);
+        dctbl1.bigTable[i] =
+            static_cast<int>((static_cast<unsigned>(x) << 8) | (l + rv));
       } else {
         dctbl1.bigTable[i] = l;
       }
@@ -306,14 +307,14 @@ public:
     int l;
     int temp;
     int code;
-    int val;
+    unsigned val;
 
     bits.fill();
     code = bits.peekBitsNoFill(14);
-    val = dctbl1.bigTable[code];
+    val = static_cast<unsigned>(dctbl1.bigTable[code]);
     if ((val & 0xff) != 0xff) {
       bits.skipBitsNoFill(val & 0xff);
-      return val >> 8;
+      return static_cast<int>(val) >> 8;
     }
 
     rv = 0;
@@ -322,7 +323,7 @@ public:
     l = val & 15;
     if (l) {
       bits.skipBitsNoFill(l);
-      rv = val >> 4;
+      rv = static_cast<int>(val) >> 4;
     } else {
       bits.skipBits(8);
       l = 8;
